Move searcher configs to dedicated directory
